refart pt2
