+++ data/5e-SRD-Conditions.json
