+++ data/5e-SRD-Equipment-Categories.json
