+++ data/5e-SRD-Equipment.json
