+++ data/5e-SRD-Magic-Items.json
