+++ data/5e-SRD-Magic-Schools.json
