+++ data/5e-SRD-Weapon-Properties.json
