+++ src/builds/magic_schools.py
@@ -1,5 +1,5 @@
 from src.utils.load_json import load_data
-from src.api.notion_call import create_page, create_database
+from src.api.notion_api import create_page, create_database
 from typing import TYPE_CHECKING, Union
 from time import sleep
 
@@ -8,7 +8,20 @@ if TYPE_CHECKING:
     from notion_client import client
 
 
-def skills_page(
+def build_magic_schools_database(logger, notion, data_directory, json_file, args):
+    magic_schools_db_id = magic_schools_db(logger, notion, args.reference_page_id)
+    magic_schools_page(
+        logger,
+        notion,
+        data_directory,
+        json_file,
+        magic_schools_db_id,
+        args.start_range,
+        args.end_range,
+    )
+
+
+def magic_schools_page(
     logger: "logging.Logger",
     notion: "client",
     data_directory: str,
@@ -18,7 +31,7 @@ def skills_page(
     json_file: str,
 ) -> None:
     """This generates the api calls needed for Notion. This parses the JSON and build the markdown body for the API call.
-    It iterates through each skills in the json depending on params.
+    It iterates through each magic_schools in the json depending on params.
 
     Args:
         logger (logging.Logger): Logging object
@@ -28,28 +41,28 @@ def skills_page(
         start (int): If you want to only capture a range specify the start
         end (Union[None, int]): If you want to only capture a range specify the end
     """
-    # == Get skills Data
-    skills_data = load_data(logger, data_directory, json_file)
+    # == Get magic_schools Data
+    magic_schools_data = load_data(logger, data_directory, json_file)
 
-    # == Apply range to skills data
-    if end is None or end > len(skills_data):
-        end = len(skills_data)
+    # == Apply range to magic_schools data
+    if end is None or end > len(magic_schools_data):
+        end = len(magic_schools_data)
 
-    # == Iterates through the specified range of the skills JSON
+    # == Iterates through the specified range of the magic_schools JSON
     for index in range(start, end):
-        selected_skill = skills_data[index]
+        schools = magic_schools_data[index]
 
         logger.info(
-            f"Building Markdown for skills -- {selected_skill['name']} -- Index -- {index} --"
+            f"Building Markdown for magic_schools -- {schools['name']} -- Index -- {index} --"
         )
 
-        # == Building markdown properties from _skills class
+        # == Building markdown properties from _magic_schools class
         markdown_properties = {
             "Name": {
                 "title": [
                     {
                         "type": "text",
-                        "text": {"content": selected_skill["name"]},
+                        "text": {"content": schools["name"]},
                     }
                 ]
             },
@@ -57,9 +70,7 @@ def skills_page(
                 "rich_text": [
                     {
                         "type": "text",
-                        "text": {
-                            "content": "".join(mn for mn in selected_skill["desc"])
-                        },
+                        "text": {"content": schools["desc"]},
                     }
                 ]
             },
@@ -68,8 +79,8 @@ def skills_page(
         # == Ensure children_properties list is empty
         children_properties = []
 
-        # == Building markdown for skills
-        children_properties = build_skills_markdown(selected_skill)
+        # == Building markdown for magic_schools
+        children_properties = build_magic_schools_markdown(schools)
 
         # == Sending api call
         # ==========
@@ -80,7 +91,9 @@ def skills_page(
         sleep(0.5)
 
 
-def skills_db(logger: "logging.Logger", notion: "client", database_id: str) -> str:
+def magic_schools_db(
+    logger: "logging.Logger", notion: "client", database_id: str
+) -> str:
     """This generates the api calls needed for Notion. This just builds the empty database page with the required options.
 
     Args:
@@ -93,7 +106,7 @@ def skills_db(logger: "logging.Logger", notion: "client", database_id: str) -> s
     """
 
     # == Database Name
-    database_name = "Skills"
+    database_name = "Magic Schools"
 
     # == Building markdown database properties
     database_weapon_properties = {
@@ -106,8 +119,8 @@ def skills_db(logger: "logging.Logger", notion: "client", database_id: str) -> s
     )
 
 
-def build_skills_markdown(skills_data: object) -> list:
-    from src.markdown.children_md import (
+def build_magic_schools_markdown(magic_schools_data: object) -> list:
+    from src.builds.children_md import (
         add_paragraph,
         add_section_heading,
         add_divider,
@@ -122,8 +135,8 @@ def build_skills_markdown(skills_data: object) -> list:
 
     # == Adding header at the top
     # ==========
-    add_section_heading(markdown_children, f"{skills_data['name']}", level=1)
+    add_section_heading(markdown_children, f"{magic_schools_data['name']}", level=1)
     add_divider(markdown_children)
-    add_paragraph(markdown_children, "".join(desc for desc in skills_data["desc"]))
+    add_paragraph(markdown_children, magic_schools_data["desc"])
 
     return markdown_children
