+++ src/builds/skills.py
@@ -1,6 +1,5 @@
-from src.markdown.weapons_properties_md import build_weapons_properties_markdown
 from src.utils.load_json import load_data
-from src.api.notion_call import create_page, create_database
+from src.api.notion_api import create_page, create_database
 from typing import TYPE_CHECKING, Union
 from time import sleep
 
@@ -9,7 +8,20 @@ if TYPE_CHECKING:
     from notion_client import client
 
 
-def weapons_properties_page(
+def build_skills_database(logger, notion, data_directory, json_file, args):
+    skills_db_id = skills_db(logger, notion, args.reference_page_id)
+    skills_page(
+        logger,
+        notion,
+        data_directory,
+        json_file,
+        skills_db_id,
+        args.start_range,
+        args.end_range,
+    )
+
+
+def skills_page(
     logger: "logging.Logger",
     notion: "client",
     data_directory: str,
@@ -19,7 +31,7 @@ def weapons_properties_page(
     json_file: str,
 ) -> None:
     """This generates the api calls needed for Notion. This parses the JSON and build the markdown body for the API call.
-    It iterates through each weapon properties in the json depending on params.
+    It iterates through each skills in the json depending on params.
 
     Args:
         logger (logging.Logger): Logging object
@@ -29,28 +41,28 @@ def weapons_properties_page(
         start (int): If you want to only capture a range specify the start
         end (Union[None, int]): If you want to only capture a range specify the end
     """
-    # == Get weapon properties Data
-    weapons_properties_data = load_data(logger, data_directory, json_file)
+    # == Get skills Data
+    skills_data = load_data(logger, data_directory, json_file)
 
-    # == Apply range to weapon properties data
-    if end is None or end > len(weapons_properties_data):
-        end = len(weapons_properties_data)
+    # == Apply range to skills data
+    if end is None or end > len(skills_data):
+        end = len(skills_data)
 
-    # == Iterates through the specified range of the weapon properties JSON
+    # == Iterates through the specified range of the skills JSON
     for index in range(start, end):
-        selected_prop = weapons_properties_data[index]
+        selected_skill = skills_data[index]
 
         logger.info(
-            f"Building Markdown for weapon properties -- {selected_prop['name']} -- Index -- {index} --"
+            f"Building Markdown for skills -- {selected_skill['name']} -- Index -- {index} --"
         )
 
-        # == Building markdown properties from _weapon properties class
+        # == Building markdown properties from _skills class
         markdown_properties = {
             "Name": {
                 "title": [
                     {
                         "type": "text",
-                        "text": {"content": selected_prop["name"]},
+                        "text": {"content": selected_skill["name"]},
                     }
                 ]
             },
@@ -59,7 +71,7 @@ def weapons_properties_page(
                     {
                         "type": "text",
                         "text": {
-                            "content": "".join(mn for mn in selected_prop["desc"])
+                            "content": "".join(mn for mn in selected_skill["desc"])
                         },
                     }
                 ]
@@ -69,8 +81,8 @@ def weapons_properties_page(
         # == Ensure children_properties list is empty
         children_properties = []
 
-        # == Building markdown for weapon properties
-        children_properties = build_weapons_properties_markdown(selected_prop)
+        # == Building markdown for skills
+        children_properties = build_skills_markdown(selected_skill)
 
         # == Sending api call
         # ==========
@@ -81,9 +93,7 @@ def weapons_properties_page(
         sleep(0.5)
 
 
-def weapons_properties_db(
-    logger: "logging.Logger", notion: "client", database_id: str
-) -> str:
+def skills_db(logger: "logging.Logger", notion: "client", database_id: str) -> str:
     """This generates the api calls needed for Notion. This just builds the empty database page with the required options.
 
     Args:
@@ -96,7 +106,7 @@ def weapons_properties_db(
     """
 
     # == Database Name
-    database_name = "Weapon Properties"
+    database_name = "Skills"
 
     # == Building markdown database properties
     database_weapon_properties = {
@@ -107,3 +117,26 @@ def weapons_properties_db(
     return create_database(
         logger, notion, database_id, database_name, database_weapon_properties
     )
+
+
+def build_skills_markdown(skills_data: object) -> list:
+    from src.builds.children_md import (
+        add_paragraph,
+        add_section_heading,
+        add_divider,
+    )
+    # == This is all of the building of the api call for
+    # == the markdown body
+    # =======================================================
+
+    # == Initializing the markdown children list
+    # ==========
+    markdown_children = []
+
+    # == Adding header at the top
+    # ==========
+    add_section_heading(markdown_children, f"{skills_data['name']}", level=1)
+    add_divider(markdown_children)
+    add_paragraph(markdown_children, "".join(desc for desc in skills_data["desc"]))
+
+    return markdown_children
