+++ src/builds/weapons.py
@@ -1,7 +1,6 @@
 from src.classes.equipment_class import _equipment
-from src.markdown.weapons_md import build_weapon_markdown
 from src.utils.load_json import load_data
-from src.api.notion_call import create_page, create_database
+from src.api.notion_api import create_page, create_database
 from typing import TYPE_CHECKING, Union
 from time import sleep
 
@@ -10,10 +9,24 @@ if TYPE_CHECKING:
     from notion_client import client
 
 
+def build_weapons_database(logger, notion, data_directory, json_file, args):
+    weapons_db_id = weapons_db(logger, notion, args.database_id)
+    weapons_page(
+        logger,
+        notion,
+        data_directory,
+        json_file,
+        weapons_db_id,
+        args.start_range,
+        args.end_range,
+    )
+
+
 def weapons_page(
     logger: "logging.Logger",
     notion: "client",
     data_directory: str,
+    json_file: str,
     database_id: str,
     start: int,
     end: Union[None, int],
@@ -30,7 +43,7 @@ def weapons_page(
         end (Union[None, int]): If you want to only capture a range specify the end
     """
     # == Get equipment Data
-    equipment_data = load_data(logger, data_directory, "5e-SRD-Equipment.json")
+    equipment_data = load_data(logger, data_directory, json_file)
 
     # == Apply range to equipment data
     if end is None or end > len(equipment_data):
@@ -221,3 +234,51 @@ def weapons_db(logger: "logging.Logger", notion: "client", database_id: str) -> 
     return create_database(
         logger, notion, database_id, database_name, database_weapon_properties
     )
+
+
+def build_weapon_markdown(equipment: object) -> list:
+    from src.builds.children_md import (
+        add_paragraph,
+        add_section_heading,
+        add_table,
+        add_divider,
+    )
+    # == This is all of the building of the api call for
+    # == the markdown body
+    # =======================================================
+
+    # == Initializing the markdown children list
+    # ==========
+    markdown_children = []
+
+    # == Adding header at the top
+    # ==========
+    add_section_heading(markdown_children, f"{equipment.name}", level=1)
+
+    headers = [
+        f"Type: {equipment.category_range}",
+        f"Cost: {equipment.cost['quantity']} {equipment.cost['unit']}",
+        f"Weight: {equipment.get_weight()}",
+    ]
+    add_table(markdown_children, headers)
+    add_divider(markdown_children)
+
+    if equipment.special:
+        for special in equipment.special:
+            add_paragraph(markdown_children, special)
+        add_divider(markdown_children)
+
+    # == Attributes
+    # ==========
+    stats_table_headers = ["Name", "Cost", "Damage", "Weight", "Properties", "Range"]
+    stats_table_row = [
+        f"{equipment.name}",
+        f"{equipment.cost['quantity']} {equipment.cost['unit']}",
+        f"{equipment.get_damage_dice()}",
+        f"{equipment.weight} lbs",
+        f"{" - ".join(prop for prop in equipment.get_properties())}",
+        f"{equipment.get_range()}",
+    ]
+    add_table(markdown_children, stats_table_headers, [stats_table_row])
+
+    return markdown_children
